simplify, adapt moe to smt design
