+++ smt/extensions/__init__.py
@@ -1,1 +1,2 @@
 from .vfm import VFM
+from .moe import MOE
