--- tests/N1_D2_100.csv
@@ -1,100 +1,0 @@
--0.2171295317234061,-0.088383614093828555,0.30551314581723465
--0.34802196980533961,0.51087428184302697,0.85889625164836658
--0.70649332712991053,-0.96198440280050068,1.6684777299304112
--0.39060358096310366,-0.23843203368487775,0.62903561464798141
-0.8239071806880276,-0.94789454748849988,1.7718017281765275
-0.36076334142557864,-0.92853986590782678,1.2893032073334054
-0.6177472625666387,0.69297687531549723,1.3107241378821359
-0.939626133469182,0.2930431259882833,1.2326692594574653
-0.55665265001665287,-0.6375526939577546,1.1942053439744074
--0.12002358058145268,-0.2488517657927436,0.36887534637419628
-0.40969703376386746,0.88534506017644854,1.295042093940316
-0.81526587832521247,0.017806407006529357,0.83307228533174182
-0.87112435207764793,-0.77475089406146569,1.6458752461391137
--0.85389200530487641,0.78805340469879659,1.641945410003673
--0.58671487311636439,0.59933245636306109,1.1860473294794254
--0.55169579334675056,-0.11450319390926578,0.66619898725601634
--0.93536889769957632,-0.57946486767365379,1.51483376537323
-0.35817923674807028,0.11953617324162158,0.47771540998969186
-0.24569885283321891,-0.42834698789683212,0.67404584073005103
--0.99118774349100713,-0.80817558995282812,1.7993633334438353
--0.8762301325274533,0.98303493471923176,1.8592650672466851
--0.46619698549801614,0.70399757659767492,1.1701945620956911
-0.57590313601168153,-0.4781414615770696,1.0540445975887511
-0.11986755375914582,-0.38512132953421063,0.50498888329335645
-0.65384472589057263,-0.15994330301116111,0.81378802890173374
-0.8590114690390287,0.66853683568273214,1.5275483047217608
-0.71369692004819973,0.5319561510922115,1.2456530711404112
--0.25801778485094806,0.62267203361391288,0.88068981846486094
-0.52897986218433335,-0.85997567225756466,1.388955534441898
--0.74245651045296368,0.96964928134775463,1.7121057918007183
-0.21756506890722704,0.42213577997665341,0.63970084888388046
-0.33099545134796982,0.77076401817365969,1.1017594695216295
--0.018990163332942567,0.54990694354928737,0.56889710688222994
-0.38876001958760908,-0.73239210293952062,1.1211521225271297
--0.78475260631327193,0.33948579122505818,1.1242383975383301
-0.89805582481840585,0.81142535046300557,1.7094811752814114
-0.68042875673974446,-0.90497499359262945,1.5854037503323739
-0.072210367193327629,-0.19335385716456399,0.26556422435789162
--0.49846118315189214,-0.65313644825446993,1.1515976314063621
-0.47353965428164924,0.60805540811407677,1.081595062395726
-0.78404378785632023,-0.27815861792894048,1.0622024057852606
--0.37858299940547879,0.1633978331592878,0.54198083256476659
--0.73813785174843227,-0.54303402546535495,1.2811718772137872
-0.75853931348371817,0.25519235376209304,1.0137316672458112
-0.95661382370991843,0.07362543652524467,1.0302392602351631
--0.067308615750971024,0.40825657606875798,0.47556519181972901
--0.66043301637323015,-0.79192954155672002,1.4523625579299502
-0.48068096367491719,0.20773141941960449,0.68841238309452168
--0.94515082958006968,-0.3033389295522303,1.2484897591322999
--0.52448274340231582,-0.33735381699730249,0.86183656039961831
--0.88679548999149294,0.18444721166534306,1.071242701656836
-0.034819680988006807,0.15001452125601311,0.18483420224401992
-0.45907348165106243,-0.35694144541387163,0.81601492706493406
-0.99997072636529771,-0.59702313370265081,1.5969938600679485
--0.29692213518762822,0.86343522199075307,1.1603573571783814
--0.63980260270483957,0.75156698480808282,1.3913695875129224
-0.052141396806893558,-0.83560728019868014,0.88774867700557369
-0.30819905693525218,-0.06822044179997655,0.37641949873522873
--0.65913513792841338,0.23562575937963803,0.89476089730805142
--0.17906915643654353,0.27630904870951123,0.45537820514605476
-0.90503305719820726,-0.13048003452772539,1.0355130917259325
-0.16281310611942956,-0.99518029881951564,1.1579934049389453
--0.18960975105689459,0.48762754992984148,0.67723730098673607
--0.82425644182925217,0.56637651230505615,1.3906329541343083
--0.6039635112641828,0.46099429605946707,1.0649578073236499
--0.3066572505917543,-0.51280333433500425,0.81946058492675855
-0.50504171632350392,0.04774254940168432,0.55278426572518824
--0.45503634836036866,-0.89425321750871534,1.349289565869084
-0.22706337073062377,0.65896225181391999,0.88602562254454376
--0.4303775478922095,0.34406513943735462,0.77444268732956412
-0.003808093992982764,0.72524666779235147,0.72905476178533424
--0.11354172279770336,-0.45604837922812758,0.56959010202583094
--0.051295506244508182,-0.059092320536612064,0.11038782678112025
-0.27277817449152408,-0.20267846667141798,0.47545664116294206
-0.97713718001253835,-0.36774277760366814,1.3448799576162065
--0.26943655753217377,-0.87547890250980576,1.1449154600419795
-0.42027061226127693,0.44059258908536814,0.86086320134664507
--0.097264724652966095,0.93327721891985282,1.0305419435728189
-0.15151123930637556,0.83512582702191485,0.9866370663282904
--0.22644602055671736,-0.68150944405386138,0.90795546461057874
--0.80027397548283341,-0.66545731565192878,1.4657312911347622
--0.69605925931698087,0.096010181011267903,0.79206944032824877
-0.6653473418114233,0.37058105445955913,1.0359283962709824
--0.91469846617131745,-0.017057645101543839,0.93175611127286129
-0.13913666235124489,-0.034159840409841991,0.17329650276108688
-0.62767635339739281,-0.29886491732936804,0.92654127072676085
-0.77186645660434983,-0.52080620997712423,1.2926726665814741
--0.76540934667313365,-0.16910491794004057,0.93451426461317422
-0.73371217696579216,-0.7448295473367097,1.4785417243025019
--0.96308206990120349,0.38257774859530902,1.3456598184965125
-0.29270750903343945,0.94690867996171524,1.2396161889951547
--0.4009581378724677,-0.40094135414272458,0.80189949201519228
--0.15994146577468016,0.12941467177697019,0.28935613755165035
--0.56757331628877039,-0.49182685269150639,1.0594001689802768
--0.33300301165109891,0.030985995775247854,0.36398900742634677
-0.58504141317114189,0.85210031458929247,1.4371417277604344
-0.1876313091458901,-0.70718747936032567,0.89481878850621577
-0.087992812492303285,0.30892541816451291,0.3969182306568162
--0.020623332880659762,-0.60000974029009435,0.62063307317075411
--0.5008173993677516,0.91421470247246295,1.4150321018402146
